Crate which

which

A Rust equivalent of Unix command which(1).

Example:

To find which rustc executable binary is using:

use which::which;
use std::path::PathBuf;

let result = which("rustc").unwrap();
assert_eq!(result, PathBuf::from("/usr/bin/rustc"));

Structs

CanonicalPath

An owned, immutable wrapper around a PathBuf containing the canonical path of an executable.

Path

An owned, immutable wrapper around a PathBuf containing the path of an executable.

WhichConfig

A wrapper containing all functionality in this crate.

Enums

Error

Functions

which

Find an executable binary’s path by name.

which_all

Find all binaries with binary_name using cwd to resolve relative paths.

which_all_global

Find all binaries with binary_name ignoring cwd.

which_global

Find an executable binary’s path by name, ignoring cwd.

which_in

Find binary_name in the path list paths, using cwd to resolve relative paths.

which_in_all

Find all binaries with binary_name in the path list paths, using cwd to resolve relative paths.

which_in_global

Find all binaries with binary_name in the path list paths, ignoring cwd.

Type Definitions

Result