Struct which::Path

pub struct Path { /* private fields */ }

An owned, immutable wrapper around a PathBuf containing the path of an executable.

The constructed PathBuf is the output of which or which_in, but which::Path has the advantage of being a type distinct from std::path::Path and std::path::PathBuf.

It can be beneficial to use which::Path instead of std::path::Path when you want the type system to enforce the need for a path that exists and points to a binary that is executable.

Since which::Path implements Deref for std::path::Path, all methods on &std::path::Path are also available to &which::Path values.

Implementations

impl Path

pub fn new<T: AsRef<OsStr>>(binary_name: T) -> Result<Path>

Returns the path of an executable binary by name.

This calls which and maps the result into a Path.

pub fn all<T: AsRef<OsStr>>( binary_name: T, ) -> Result<impl Iterator<Item = Path>>

Returns the paths of all executable binaries by a name.

this calls which_all and maps the results into Paths.

pub fn new_in<T, U, V>(binary_name: T, paths: Option<U>, cwd: V) -> Result<Path>
where T: AsRef<OsStr>, U: AsRef<OsStr>, V: AsRef<Path>,

Returns the path of an executable binary by name in the path list paths and using the current working directory cwd to resolve relative paths.

This calls which_in and maps the result into a Path.

pub fn all_in<T, U, V>( binary_name: T, paths: Option<U>, cwd: V, ) -> Result<impl Iterator<Item = Path>>
where T: AsRef<OsStr>, U: AsRef<OsStr>, V: AsRef<Path>,

Returns all paths of an executable binary by name in the path list paths and using the current working directory cwd to resolve relative paths.

This calls which_in_all and maps the results into a Path.

pub fn as_path(&self) -> &Path

Returns a reference to a std::path::Path.

pub fn into_path_buf(self) -> PathBuf

Consumes the which::Path, yielding its underlying std::path::PathBuf.