which

Trait NonFatalErrorHandler

pub trait NonFatalErrorHandler {
    // Required method
    fn handle(&mut self, e: NonFatalError);
}

Defines what should happen when a nonfatal error is encountered. A nonfatal error may represent a problem, but it doesn’t necessarily require which to stop its search.

This trait is implemented for any closure or function that takes a single argument which is a NonFatalError. You may also implement it for your own types.

Required Methods

fn handle(&mut self, e: NonFatalError)